requirements.txt
boto3==1.17.84
celery==5.1.0
dj-database-url==0.5.0
django==3.2.4
django-configurations==2.2
django-extensions==3.1.3
django-public-admin==0.0.5
django-simple-history==3.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
dramatiq[rabbitmq,watch]==1.11.0
gunicorn==20.1.0
jinja2==3.0.1
psycopg2-binary==2.8.6
python-dotenv==0.17.1
schematics==2.1.0
scrapy==2.5.0
sentry-dramatiq==0.3.2
sentry-sdk==1.1.0
spidermon==1.15.0
tika==1.24
whitenoise==5.2.0
xlrd==2.0.1
